// frontend/src/api/model/alertThresholdMetric.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DevOps Dashboard API
 * 运维监控仪表盘一期 API 规范（Spec-Driven Development）
 * OpenAPI spec version: 1.0.0
 */

export type AlertThresholdMetric = typeof AlertThresholdMetric[keyof typeof AlertThresholdMetric];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AlertThresholdMetric = {
  cpu: 'cpu',
  memory: 'memory',
  disk: 'disk',
} as const;